+++ nautobot_netbox_importer/tests/test_import.py
@@ -4,17 +4,19 @@ import os
 from packaging import version
 
 from django.contrib.contenttypes.models import ContentType
-from django.core.exceptions import ObjectDoesNotExist
+from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.test import TestCase
 import yaml
 
+from nautobot.dcim.models import Rack
 from nautobot.extras.models import ObjectChange, ChangeLoggedModel
 
 from nautobot_netbox_importer.management.commands.import_netbox_json import Command
 
 
-NETBOX_DATA_FILE = os.path.join(os.path.dirname(__file__), "fixtures", "netbox_dump.json")
-NAUTOBOT_DATA_FILE = os.path.join(os.path.dirname(__file__), "fixtures", "nautobot_expectations.yaml")
+FIXTURES_DIR = os.path.join(os.path.dirname(__file__), "fixtures")
+NETBOX_DATA_FILE = os.path.join(FIXTURES_DIR, "netbox_dump.json")
+NAUTOBOT_DATA_FILE = os.path.join(FIXTURES_DIR, "nautobot_expectations.yaml")
 
 
 class TestImport(TestCase):
@@ -26,7 +28,12 @@ class TestImport(TestCase):
     def setUpTestData(cls) -> None:
         """One-time setup function called before running the test functions in this class."""
         with open(NETBOX_DATA_FILE, "r") as file_handle:
-            Command().handle(json_file=file_handle, netbox_version=version.parse("2.10.4"), verbosity=0)
+            Command().handle(
+                json_file=file_handle,
+                netbox_version=version.parse("2.10.4"),
+                verbosity=0,
+                bypass_data_validation=False,
+            )
         # TODO check stdout/stderr for errors and such
         with open(NAUTOBOT_DATA_FILE, "r") as handle:
             cls.nautobot_data = yaml.safe_load(handle)
@@ -99,6 +106,48 @@ class TestImport(TestCase):
     def test_resync_without_changes_correctness(self):
         """Resync (with no changes to the source data) and verify that data is still correct."""
         with open(NETBOX_DATA_FILE, "r") as file_handle:
-            Command().handle(json_file=file_handle, netbox_version=version.parse("2.10.4"), verbosity=0)
+            Command().handle(
+                json_file=file_handle,
+                netbox_version=version.parse("2.10.4"),
+                verbosity=0,
+                bypass_data_validation=False,
+            )
         # TODO check stdout/stderr for errors and such
         self.test_imported_data_correctness()
+
+
+class TestImportInvalid(TestCase):
+    """Test the import handling of invalid source data, with and without the --bypass-data-validation flag."""
+
+    maxDiff = None
+
+    def test_invalid_rack(self):
+        """Import a Rack whose assigned Site doesn't match the Site of its assigned RackGroup."""
+        netbox_data_file = os.path.join(FIXTURES_DIR, "netbox_dump_invalid_rack.json")
+
+        with open(netbox_data_file, "r") as file_handle:
+            Command().handle(
+                json_file=file_handle,
+                netbox_version=version.parse("2.10.3"),
+                verbosity=0,
+                bypass_data_validation=False,
+            )
+        # TODO check stdout/stderr for expected error messages
+
+        # Rack should NOT have been created since it failed model validation during clean()
+        self.assertEqual(0, Rack.objects.count(), "Rack was created even though it failed model validation??")
+
+        # Try again, this time with the flag
+        with open(netbox_data_file, "r") as file_handle:
+            Command().handle(
+                json_file=file_handle,
+                netbox_version=version.parse("2.10.3"),
+                verbosity=0,
+                bypass_data_validation=True,
+            )
+        # TODO check stdout/stderr for expected warning message but no errors
+
+        self.assertEqual(1, Rack.objects.count(), "Rack was not created despite bypassing model validation??")
+        rack = Rack.objects.get(name="invalid-site-group-site-mismatch")
+        with self.assertRaises(ValidationError):
+            rack.clean()